Add Pass.Manager.Resources project Add compression tests Add Html.ActionLinkAjaxExt() Add FileExtensions and tests Add missed fields in PassBeacon, PassLocation Add PassTemplateOnlineService and PassTemplateOnlineController to register PassContentTemplate in Pass.Container (not finished)

+++ MobileWallet/Pass.Container/Pass.Container.Core/Entities/Templates/GeneralPassTemplate/BeaconDetails.cs
@@ -22,11 +22,11 @@ namespace Pass.Container.Core.Entities.Templates.GeneralPassTemplate
 
         [XmlElement(ElementName = "major")]
         [JsonProperty(PropertyName = "major", NullValueHandling = NullValueHandling.Ignore)]
-        public Int16? Major { get; set; }
+        public int? Major { get; set; }
 
         [XmlElement(ElementName = "minor")]
         [JsonProperty(PropertyName = "minor", NullValueHandling = NullValueHandling.Ignore)]
-        public Int16? Minor { get; set; }
+        public int? Minor { get; set; }
 
         [XmlElement(ElementName = "proximityUUID")]
         [JsonProperty(PropertyName = "proximityUUID", Required = Required.Always)]
